tools: imageconfigvalidator: refactor as a lib instead of main package
